refator trpc usage in app router
