--- apps/web/app/_trpc/HydrateClient.tsx
@@ -1,8 +1,0 @@
-"use client";
-
-import { createHydrateClient } from "app/_trpc/createHydrateClient";
-import superjson from "superjson";
-
-export const HydrateClient = createHydrateClient({
-  transformer: superjson,
-});
