--- apps/web/app/_trpc/serverClient.ts
@@ -1,4 +1,0 @@
-import type { TRPCContext } from "@calcom/trpc/server/createContext";
-import { appRouter } from "@calcom/trpc/server/routers/_app";
-
-export const getServerCaller = (ctx: TRPCContext) => appRouter.createCaller(ctx);
