+++ apps/web/app/providers.tsx
@@ -1,8 +1,10 @@
 "use client";
 
-import { type DehydratedState } from "@tanstack/react-query";
+import { queryClient } from "app/_trpc/query-client";
+import { trpcClient } from "app/_trpc/trpc-client";
 import { TrpcProvider } from "app/_trpc/trpc-provider";
 import { SessionProvider } from "next-auth/react";
+import { useEffect } from "react";
 import CacheProvider from "react-inlinesvg/provider";
 
 import { WebPushProvider } from "@calcom/features/notifications/WebPushContext";
@@ -12,14 +14,33 @@ import PlainChat from "@lib/plain/dynamicProvider";
 
 type ProvidersProps = {
   children: React.ReactNode;
-  dehydratedState: DehydratedState | undefined;
 };
-export function Providers({ children, dehydratedState }: ProvidersProps) {
+export function Providers({ children }: ProvidersProps) {
   const isBookingPage = useIsBookingPage();
+
+  // Prefetching
+  useEffect(() => {
+    queryClient.prefetchQuery({
+      queryKey: ["viewer.features.map"],
+      queryFn: () => trpcClient.viewer.features.map.query(),
+    });
+    queryClient.prefetchQuery({
+      queryKey: ["viewer.teams.hasTeamPlan"],
+      queryFn: () => trpcClient.viewer.teams.hasTeamPlan.query(),
+    });
+    queryClient.prefetchQuery({
+      queryKey: ["viewer.public.session"],
+      queryFn: () => trpcClient.viewer.public.session.query(),
+    });
+    queryClient.prefetchQuery({
+      queryKey: ["viewer.me.get"],
+      queryFn: () => trpcClient.viewer.me.get.query(),
+    });
+  }, []);
 
   return (
     <SessionProvider>
-      <TrpcProvider dehydratedState={dehydratedState}>
+      <TrpcProvider>
         {!isBookingPage ? <PlainChat /> : null}
         {/* @ts-expect-error FIXME remove this comment when upgrading typescript to v5 */}
         <CacheProvider>
